refactor(ui): extract feed/filter.ts — filterByType/filterByQuery/computeSignature
